feat(eventuser): security changes for event
takes the user id by the token and does not let you delete or update other users' events

+++ src/modules/event/services/eventUser/DeleteEventUserService.ts
@@ -3,6 +3,7 @@ import AppError from '@shared/infra/http/errors/AppError';
 import IEventUserRepository from '../../repositories/IEventUserRepository';
 
 interface IRequest {
+  user_id: string;
   eventUser_id: string;
 }
 
@@ -13,11 +14,15 @@ export default class DeleteEventUserService {
     private eventUserRepository: IEventUserRepository,
   ) {}
 
-  public async execute({ eventUser_id }: IRequest): Promise<void> {
+  public async execute({ user_id, eventUser_id }: IRequest): Promise<void> {
     const eventUser = await this.eventUserRepository.findById(eventUser_id);
 
     if (!eventUser) {
       throw new AppError('Event does not exists.', 404);
+    }
+
+    if (user_id !== eventUser.user_id) {
+      throw new AppError('This event does not belong to this user.', 404);
     }
 
     await this.eventUserRepository.delete(eventUser.id);
